rename/split files/funcs in inference/latent
