+++ docs/api_reference/developer/inference/latent_path_builder.md
@@ -4,12 +4,12 @@ Developer-facing reference for latent-path construction and scoring.
 
 The latent-state refactor separates:
 
-- `dynestyx.inference.latent.builder`: handler logic and NumPyro-site
+- `dynestyx.inference.latent.base`: handler logic and NumPyro-site
   registration
-- `dynestyx.inference.latent.state_path`: pure-JAX path assembly helpers
-- `dynestyx.inference.latent.trajectory_log_probs`: pure-JAX joint trajectory
-  scoring helpers
+- `dynestyx.inference.latent.metadata`: latent layout / indexing / canonicalization
+- `dynestyx.inference.latent.assembly`: pure-JAX state-path reconstruction
+- `dynestyx.inference.latent.scoring`: pure-JAX joint trajectory scoring
 
-::: dynestyx.inference.latent.builder
+::: dynestyx.inference.latent.base
     options:
       filters: []
