+++ dynestyx/inference/latent/__init__.py
@@ -1,1 +1,39 @@
-"""Latent-path subpackage for state construction and scoring."""
+"""Latent-path construction, assembly, and scoring helpers."""
+
+from dynestyx.inference.latent.assembly import (
+    AssembledStateTrajectory,
+    assemble_dirac_state_path,
+    assemble_state_path,
+    default_ode_diffeqsolve_settings,
+)
+from dynestyx.inference.latent.base import LatentPathBuilder
+from dynestyx.inference.latent.metadata import (
+    DiracLatentMetadata,
+    canonicalize_dirac_state_path_params,
+    canonicalize_state_path_params,
+    fully_observed_dirac_state_path_param_metadata,
+    infer_dirac_state_path_param_metadata,
+    infer_state_path_param_times,
+    prepare_dirac_state_path_metadata,
+)
+from dynestyx.inference.latent.scoring import (
+    TrajectoryLogProbTerms,
+    compute_trajectory_log_prob_terms,
+)
+
+__all__ = [
+    "AssembledStateTrajectory",
+    "DiracLatentMetadata",
+    "LatentPathBuilder",
+    "TrajectoryLogProbTerms",
+    "assemble_dirac_state_path",
+    "assemble_state_path",
+    "canonicalize_dirac_state_path_params",
+    "canonicalize_state_path_params",
+    "compute_trajectory_log_prob_terms",
+    "default_ode_diffeqsolve_settings",
+    "fully_observed_dirac_state_path_param_metadata",
+    "infer_dirac_state_path_param_metadata",
+    "infer_state_path_param_times",
+    "prepare_dirac_state_path_metadata",
+]
